+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+**Added**
+
+* Added configuration options `enableUnifiedRedirectDomain` and `unifiedRedirectDomain` to allow using one unified domain for the IdP redirect to the shop. (copied from 8.1.0)
+
 # 9.0.0
 
 **Added**